Hotel Reservation Prediction/pyproject.toml
[project]
name = "hotel-reservation-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pandas", "numpy", "google-cloud-storage", "scikit-learn", "pyyaml",
    "matplotlib", "seaborn", "statsmodels", "imbalanced-learn", "ipykernel",
    "xgboost", "lightgbm","mlflow", "setuptools"]

